network: add support of configurable bpf clsact filter program for container interfaces
* Adds net_limit_soft property to a container, that can be changed at runtime.

 * Adds support for `network { network_limit_soft_bpf_elf_path: "/path/to/bpf-elf.o" }` option to portod.conf.

 * Adds a configured bpf program to container interfaces.

 * Adds setting up of a bpf map named "netlimit_soft_map" and writing net_limit_soft
   property of a container to a network inode key in the map.
02c107343d3e4f0d09cec3141945b4303bcef2e8

+++ src/util/netlink.cpp
@@ -1916,7 +1916,7 @@ TError TNlBpfFilter::Create(const TNl &nl) {
     tchdr.tcm_ifindex = Index;
     tchdr.tcm_handle = 0;
     tchdr.tcm_parent = TC_H_MAKE(TC_H_CLSACT, Egress ? TC_H_MIN_EGRESS : TC_H_MIN_INGRESS);
-    tchdr.tcm_info = TC_H_MAKE(FilterPrio << 16, htons(ETH_P_ALL));
+    tchdr.tcm_info = TC_H_MAKE(((uint32_t)FilterPrio) << 16, htons(ETH_P_ALL));
 
     msg = nlmsg_alloc_simple(RTM_NEWTFILTER, NLM_F_EXCL | NLM_F_CREATE);
     if (!msg)
